Feat : test yml 4
Former-commit-id: c9da29448e3f3625954c43ce0bf635e854cba64e

+++ backend/src/main/java/com/a503/churros/feign/news/NewsFeign.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface NewsFeign {
 
     @GetMapping("/{userId}")
-    List<Long> getRecomList(@PathVariable("userId") long userId);
+    List<Integer> getRecomList(@PathVariable("userId") long userId);
 
     @GetMapping("/sample")
-    List<Long> getSampleList();
+    List<Integer> getSampleList();
 
 }
